+++ presentation/src/main/java/com/anytypeio/anytype/presentation/page/editor/control/ControlPanelState.kt
@@ -124,6 +124,7 @@ data class ControlPanelState(
         data class MultiSelect(
             override val isVisible: Boolean,
             val isScrollAndMoveEnabled: Boolean = false,
+            val isQuickScrollAndMoveMode: Boolean = false,
             val count: Int = 0
         ) : Toolbar()