+++ presentation/src/test/java/com/anytypeio/anytype/presentation/page/editor/EditorQuickStartingScrollAndMoveTest.kt
@@ -120,7 +120,8 @@ class EditorQuickStartingScrollAndMoveTest : EditorPresentationTestSetup() {
                 multiSelect = Toolbar.MultiSelect(
                     isVisible = true,
                     count = 1,
-                    isScrollAndMoveEnabled = true
+                    isScrollAndMoveEnabled = true,
+                    isQuickScrollAndMoveMode = true,
                 ),
                 stylingToolbar = Toolbar.Styling(isVisible = false)
             )
@@ -146,6 +147,111 @@ class EditorQuickStartingScrollAndMoveTest : EditorPresentationTestSetup() {
                         text = b.content<TXT>().text,
                         mode = BlockView.Mode.READ,
                         isSelected = true
+                    )
+                )
+            )
+        )
+    }
+
+    @Test
+    fun `should exit to editing-mode when exiting from quick SAM`() {
+
+        // SETUP
+
+        val a = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Text(
+                text = MockDataFactory.randomString(),
+                marks = emptyList(),
+                style = Block.Content.Text.Style.NUMBERED
+            )
+        )
+
+        val b = Block(
+            id = MockDataFactory.randomUuid(),
+            fields = Block.Fields.empty(),
+            children = emptyList(),
+            content = Block.Content.Text(
+                text = MockDataFactory.randomString(),
+                marks = emptyList(),
+                style = Block.Content.Text.Style.P
+            )
+        )
+
+        val page = Block(
+            id = root,
+            fields = Block.Fields(emptyMap()),
+            content = Block.Content.Smart(
+                type = Block.Content.Smart.Type.PAGE
+            ),
+            children = listOf(header.id, a.id, b.id)
+        )
+
+        val document = listOf(page, header, title, a, b)
+
+        stubOpenDocument(document)
+        stubInterceptEvents()
+
+        val vm = buildViewModel()
+
+        // TESTING
+
+        vm.onStart(root)
+
+        val controlPanelTestObserver = vm.controlPanelViewState.test()
+        val viewStateTestObserver = vm.state.test()
+
+        vm.apply {
+            onBlockFocusChanged(id = b.id, hasFocus = true)
+            onBlockToolbarBlockActionsClicked()
+            onActionMenuItemClicked(id = b.id, action = ActionItemType.SAM)
+            onExitScrollAndMoveClicked()
+        }
+
+        // VERIFYING
+
+        controlPanelTestObserver.assertValue(
+            ControlPanelState(
+                navigationToolbar = Toolbar.Navigation(isVisible = true),
+                mainToolbar = Toolbar.Main(isVisible = false),
+                mentionToolbar = Toolbar.MentionToolbar(
+                    isVisible = false,
+                    cursorCoordinate = null,
+                    mentionFilter = null,
+                    mentionFrom = null
+                ),
+                multiSelect = Toolbar.MultiSelect(
+                    isVisible = false,
+                    count = 0,
+                    isScrollAndMoveEnabled = false,
+                    isQuickScrollAndMoveMode = false,
+                ),
+                stylingToolbar = Toolbar.Styling(isVisible = false)
+            )
+        )
+
+        viewStateTestObserver.assertValue(
+            ViewState.Success(
+                listOf(
+                    BlockView.Title.Document(
+                        id = title.id,
+                        text = title.content<TXT>().text,
+                        mode = BlockView.Mode.EDIT,
+                    ),
+                    BlockView.Text.Numbered(
+                        id = a.id,
+                        text = a.content<TXT>().text,
+                        number = 1,
+                        mode = BlockView.Mode.EDIT,
+                        isSelected = false
+                    ),
+                    BlockView.Text.Paragraph(
+                        id = b.id,
+                        text = b.content<TXT>().text,
+                        mode = BlockView.Mode.EDIT,
+                        isSelected = false
                     )
                 )
             )